backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pgvector==0.2.4
insightface==0.7.3
# onnxruntime wheels differ by Python version on macOS.
# - Python 3.11+ generally needs onnxruntime 1.20+
# - Python 3.9/3.10 typically supports up to 1.19.x on PyPI
onnxruntime>=1.20.0,<2; python_version >= "3.11"
onnxruntime>=1.16.0,<1.20; python_version < "3.11"
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.26.2
boto3==1.29.7
minio==7.2.0
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pillow-heif==0.13.0
